ai_forex_sentinel/requirements_server.txt
streamlit
yfinance>=0.2.51
pandas
numpy
plotly
matplotlib
feedparser
vaderSentiment
twilio
v20
oandapyV20
google-genai
PyYAML
hmmlearn
torch
scikit-learn
requests
python-dotenv
psutil
alpaca-py
pandas-ta
